[feat] #6 Service 분리
