+++ data/src/main/java/com/jslee/data/remote/datasource/impl/GeocodingDataSourceImpl.kt
@@ -2,7 +2,7 @@ package com.jslee.data.remote.datasource.impl
 
 import com.jslee.data.remote.datasource.GeocodingDataSource
 import com.jslee.data.remote.dto.GeocodingResponse
-import com.jslee.data.remote.service.WeatherService
+import com.jslee.data.remote.service.GeocodingService
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import javax.inject.Inject
@@ -13,13 +13,13 @@ import javax.inject.Inject
  * @created 2023/05/16
  */
 internal class GeocodingDataSourceImpl @Inject constructor(
-    private val weatherService: WeatherService
+    private val geocodingService: GeocodingService
 ) : GeocodingDataSource {
     override fun getLocationName(
         query: String
     ): Flow<List<GeocodingResponse>> = flow {
         runCatching {
-            weatherService.getLocationName(query = "$query, KR")
+            geocodingService.getLocationName(query = "$query, KR")
         }.getOrDefault(emptyList())
     }
 }
