+++ data/src/main/java/com/jslee/data/remote/service/WeatherService.kt
@@ -2,6 +2,7 @@ package com.jslee.data.remote.service
 
 import com.jslee.data.BuildConfig
 import com.jslee.data.remote.dto.CurrentWeatherResponse
+import com.jslee.data.remote.dto.ForecastResponse
 import com.jslee.data.remote.dto.GeocodingResponse
 import retrofit2.http.GET
 import retrofit2.http.Query
@@ -12,14 +13,6 @@ import retrofit2.http.Query
  * @created 2023/05/16
  */
 internal interface WeatherService {
-
-    @GET("geo/1.0/direct")
-    fun getLocationName(
-        @Query("q") query: String,
-        @Query("limit") limit: Int = 10,
-        @Query("appid") apiKey: String = BuildConfig.API_KEY,
-    ): List<GeocodingResponse>
-
     @GET("data/2.5/weather")
     fun getCurrentWeather(
         @Query("lat") latitude: Double,
@@ -36,5 +29,5 @@ internal interface WeatherService {
         @Query("appid") apiKey: String = BuildConfig.API_KEY,
         @Query("lang") language: String = "kr",
         @Query("units") units: String = "metric",
-    ): CurrentWeatherResponse
+    ): ForecastResponse
 }
